+++ frontend/centero/lib/pages/call_regular.dart
@@ -34,7 +34,7 @@ class PageCallRegular extends StatelessWidget {
 
     Widget nameHeader() {
       return NameHeader(
-        child: Text('You are speaking with: Jane Doe', 
+        child: Text('You are speaking with: Marcus Bennett', 
           style: TextStyle(fontSize: 80, fontFamily: 'Josefin', fontWeight: FontWeight.bold), 
           textAlign: TextAlign.center));
     }
