+++ frontend/centero/lib/pages/call_regular_confirmation.dart
@@ -42,7 +42,7 @@ class PageCallRegularConfirmation extends StatelessWidget {
               text: 'Connecting you now to ',
               style: TextStyle(fontSize: 80, color: Colors.white, fontFamily: 'Josefin', fontWeight: FontWeight.bold),
               children: const <TextSpan>[
-                TextSpan(text: 'Mark Smith...', style: TextStyle(fontWeight: FontWeight.w800)),
+                TextSpan(text: 'Marcus Bennett...', style: TextStyle(fontWeight: FontWeight.w800)),
               ],
             ),
           ),
